requirements.txt
appdirs==1.4.4
boto3==1.24.11
botocore==1.27.11
certifi==2022.6.15
charset-normalizer==2.0.12
colored==1.4.3
cuda-python==11.7.0
cycler==0.11.0
Cython==0.29.30
easydict==1.9
flatbuffers==2.0
fonttools==4.33.3
graphsurgeon @ file:///TensorRT-8.2.4.2/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
h5py==3.7.0
idna==3.3
jmespath==1.0.0
json-lines==0.5.0
jsonlines==1.2.0
kiwisolver==1.4.3
lmdb==0.94
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
msgpack==1.0.4
msgpack-numpy==0.4.8
numpy==1.22.3
nvidia-pyindex==1.0.9
onnx==1.11.0
onnx-graphsurgeon==0.3.19
onnxruntime-gpu==1.11.1
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.1
polygraphy==0.36.2
protobuf==3.19.4
pycuda==2021.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytools==2022.1.2
pytorch-pretrained-bert==0.6.2
PyYAML==5.1.2
pyzmq==23.1.0
regex==2022.6.2
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
tabulate==0.8.9
tensorboardX==1.2
tensorpack==0.9.4
tensorrt @ file:///workspace/TensorRT-8.4.1.5-GA/tensorrt-8.4.1.5-cp38-none-linux_x86_64.whl
termcolor==1.1.0
torch==1.8.1+cu111
tqdm==4.31.1
typing_extensions==4.2.0
uff @ file:///TensorRT-8.2.4.2/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.9
